requirements.txt
sacrebleu==1.4.3
spacy
tqdm==4.42.0
tensorflow-gpu==2.1
matplotlib==3.1.1
numpy==1.16.0
jsonschema==3.2.0
pandas==0.25.3
sklearn
click
